+++ CHANGELOG.md
@@ -23,9 +23,11 @@ All Notable changes to `Backpack CRUD` will be documented in this file
 
 ## [3.3.9] - 2018-02-xx
 
+### Added
+- setLabeller() method to change how labels are made; merges #688;
+
 ## Fixed
 - double-click on create form created two entries; fixes #1229;
-
 
 
 ## [3.3.8] - 2018-02-08